+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 
-## 0.9.0 (unreleased)
+## 0.9.0 (2024-11-03)
 
 - Fixed parsing of empty series document.